requirements.txt
google-api-python-client~=2.163.0
google-auth-httplib2~=0.2.0
google-auth-oauthlib~=1.2.1
tzlocal~=5.3.1
html2text~=2025.4.15
pydantic~=2.11.7
pip~=25.2
typing_extensions~=4.15.0
attrs~=25.1.0
protobuf~=5.29.3
cryptography~=43.0.3
docutils~=0.22
filelock~=3.19.1
rsa~=4.9
pyasn1~=0.6.1
pyzmq~=26.2.1
colorama~=0.4.6
pytest~=8.0.2
cffi~=1.17.1
tornado~=6.4.2
comm~=0.2.2
parso~=0.8.4
jedi~=0.19.2
pywin32~=309
proto-plus~=1.26.0
setuptools~=76.0.0
aiohttp~=3.12.15
pyasn1_modules~=0.4.1
cachetools~=5.5.2
certifi~=2025.1.31
urllib3~=2.3.0
packaging~=24.2
requests~=2.32.3
psutil~=7.0.0
debugpy~=1.8.13
ipython~=9.0.2
traitlets~=5.14.3
greenlet~=3.2.4
SQLAlchemy~=2.0.42
decorator~=5.2.1
Pygments~=2.19.1
ipython_pygments_lexers~=1.1.1
stack-data~=0.6.3
jupyter_client~=8.6.3
prompt_toolkit~=3.0.50
wcwidth~=0.2.13
six~=1.17.0
python-dateutil~=2.9.0.post0
pyparsing~=3.2.1
httplib2~=0.22.0
oauthlib~=3.2.2
pydantic_core~=2.33.2
annotated-types~=0.7.0
rich~=14.1.0
pillow~=11.3.0
idna~=3.10
charset-normalizer~=3.4.1
asttokens~=3.0.0
executing~=2.2.0
ipykernel~=6.29.5
matplotlib-inline~=0.1.7
jupyter_core~=5.7.2
pure_eval~=0.2.3
Jinja2~=3.1.6
uritemplate~=4.1.1
platformdirs~=4.3.6
click~=8.2.1
requests-oauthlib~=2.0.0